requirements.txt
jax
jaxlib
flax
optax
hydra-core
jaxtyping
chex
wandb
numpy
scipy
pytest
rdkit
pandas
transformers
datasets
